+++ src/views/activityBar/serverView/ServerDataProvider.ts
@@ -10,7 +10,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing
 // permissions and limitations under the License.
-import { EventEmitter, ThemeColor, ThemeIcon, TreeDataProvider, TreeItem } from 'vscode';
+import { EventEmitter, TreeDataProvider, TreeItem } from 'vscode';
 import { State } from 'vscode-languageclient';
 import { checkServerStatus, isServerStatus } from '../../../commands/server/utils';
 import { EventBus } from '../../../services/EventBus';
@@ -21,6 +21,7 @@ import {
   LSP_ZENML_CLIENT_INITIALIZED,
   SERVER_STATUS_UPDATED,
 } from '../../../utils/constants';
+import { TREE_ICONS } from '../../../utils/ui-constants';
 import { LOADING_TREE_ITEMS } from '../common/LoadingTreeItem';
 import { ServerTreeItem } from './ServerTreeItems';
 
@@ -152,12 +153,9 @@ export class ServerDataProvider implements TreeDataProvider<TreeItem> {
   getTreeItem(element: TreeItem): TreeItem {
     if (element instanceof ServerTreeItem) {
       if (element.serverStatus.isConnected) {
-        element.iconPath = new ThemeIcon(
-          'vm-active',
-          new ThemeColor('gitDecoration.addedResourceForeground')
-        );
+        element.iconPath = TREE_ICONS.SERVER_CONNECTED;
       } else {
-        element.iconPath = new ThemeIcon('vm-connect');
+        element.iconPath = TREE_ICONS.SERVER_DISCONNECTED;
       }
     }
     return element;